Refactored relation property components
