+++ components/common/BoardEditor/focalboard/src/widgets/menu/RelationPropertyMenu/RelationPropertyOptions.tsx
@@ -1,8 +1,5 @@
-import styled from '@emotion/styled';
 import KeyboardArrowRightIcon from '@mui/icons-material/KeyboardArrowRight';
-import { Box, Divider, Menu, MenuItem, MenuList, Stack, Switch, TextField, Typography } from '@mui/material';
-import { bindMenu } from 'material-ui-popup-state';
-import type { PopupState } from 'material-ui-popup-state/hooks';
+import { MenuItem, MenuList, Stack, Switch, TextField, Typography } from '@mui/material';
 import { useEffect, useState } from 'react';
 
 import { useSyncRelationProperty } from 'charmClient/hooks/blocks';
@@ -10,120 +7,11 @@ import { Button } from 'components/common/Button';
 import { PageIcon } from 'components/common/PageIcon';
 import PopperPopup from 'components/common/PopperPopup';
 import { usePages } from 'hooks/usePages';
-import type { Board, IPropertyTemplate, PropertyType, RelationPropertyData } from 'lib/focalboard/board';
+import type { Board, IPropertyTemplate, RelationPropertyData } from 'lib/focalboard/board';
 
-import { LinkCharmVerseDatabase } from '../../components/viewSidebar/viewSourceOptions/components/LinkCharmVerseDatabase';
-import { useAppSelector } from '../../store/hooks';
+import { useAppSelector } from '../../../store/hooks';
 
-const StyledMenuItem = styled(MenuItem)`
-  padding: ${({ theme }) => theme.spacing(0, 1)};
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  flex-direction: row;
-`;
-
-export function RelationPropertyMenu({
-  onClick,
-  popupState,
-  relationData
-}: {
-  popupState: PopupState;
-  onClick: (property: {
-    type: PropertyType;
-    relationData?: IPropertyTemplate['relationData'];
-    name?: IPropertyTemplate['name'];
-  }) => void;
-  relationData?: IPropertyTemplate['relationData'];
-}) {
-  const bindMenuProps = bindMenu(popupState);
-  const [relationPropertyData, setRelationPropertyData] = useState<IPropertyTemplate['relationData'] | null>(
-    relationData ?? null
-  );
-  const [propertyTitle, setPropertyTitle] = useState('Relation' || '');
-  const [showSelectDatabaseMenu, setShowSelectDatabaseMenu] = useState(true);
-
-  function onClose() {
-    bindMenuProps.onClose();
-    // Wait for the menu to close before resetting the state
-    setTimeout(() => {
-      setShowSelectDatabaseMenu(false);
-      setRelationPropertyData(null);
-    }, 150);
-  }
-
-  return (
-    <Menu
-      {...bindMenuProps}
-      onClick={(e) => {
-        e.stopPropagation();
-      }}
-      onClose={onClose}
-    >
-      {showSelectDatabaseMenu ? (
-        <LinkCharmVerseDatabase
-          placeholder='Link to a database'
-          onSelectLinkedDatabase={({ sourceDatabaseId, pageTitle = 'Untitled' }) => {
-            setRelationPropertyData(
-              relationPropertyData
-                ? {
-                    ...relationPropertyData,
-                    boardId: sourceDatabaseId
-                  }
-                : {
-                    boardId: sourceDatabaseId,
-                    limit: 'single_page',
-                    relatedPropertyId: null,
-                    showOnRelatedBoard: false
-                  }
-            );
-            setPropertyTitle(`Related to ${pageTitle}`);
-            setShowSelectDatabaseMenu(false);
-          }}
-        />
-      ) : relationPropertyData ? (
-        <Box
-          sx={{
-            minWidth: 200,
-            px: 1
-          }}
-        >
-          <TextField
-            value={propertyTitle}
-            onClick={(e) => {
-              e.stopPropagation();
-            }}
-            fullWidth
-            onChange={(e) => {
-              e.stopPropagation();
-              setPropertyTitle(e.target.value);
-            }}
-            autoFocus
-          />
-          <Divider
-            sx={{
-              my: 1
-            }}
-          />
-          <RelationPropertyCreateOptions
-            onChange={setRelationPropertyData}
-            relationData={relationPropertyData}
-            onButtonClick={() => {
-              onClick({
-                type: 'relation',
-                relationData: relationPropertyData,
-                name: propertyTitle
-              });
-              onClose();
-            }}
-            disabled={propertyTitle.length === 0}
-            setShowSelectDatabaseMenu={setShowSelectDatabaseMenu}
-          />
-        </Box>
-      ) : null}
-    </Menu>
-  );
-}
+import { StyledMenuItem } from './RelationPropertyMenu';
 
 function RelationPropertyOptions({
   relationData,
